Citrus Fruit and Leaf Disease Dataset
The Citrus Fruits and Leaves dataset is a collection of images of various citrus diseases, such as black spot, decay, and huanglongbing or greening, along with healthy citrus. The dataset contains a total of 9304 images, with 4 classes of fruits and leaves. The picture is labeled with the name of the disease of each type of citrus. Datasets can be used for tasks such as machine learning, and deep learning.